+++ src/main/java/com/grupop/gestion/Servicios/PagoServicio.java
@@ -1,5 +1,6 @@
 package com.grupop.gestion.Servicios;
 
+import com.grupop.gestion.DTO.OperacionesDTO;
 import com.grupop.gestion.DTO.PagoDTO;
 import com.grupop.gestion.Entidades.Cobro;
 import com.grupop.gestion.Entidades.EntidadBase;
@@ -28,7 +29,7 @@ public class PagoServicio {
     private final EntidadBaseServicio entidadBaseServicio;
 
     @Transactional
-    public void crear(Pago dto, String fechaComprobante){
+    public void crear(Pago dto, String fechaComprobante) {
         Pago p = new Pago();
         p.setProveedor(dto.getProveedor());
         p.setFechaComprobante(LocalDate.parse(fechaComprobante));
@@ -48,62 +49,72 @@ public class PagoServicio {
     }
 
     @Transactional
-    public void actualizar(Pago dto, String fechaComprobante){
-        Pago p = pagoRepo.findById(dto.getId()).get();
-        Long idFormaDePagoAnterior;
-        try{
-            idFormaDePagoAnterior = p.getFormaDePago().getId();
-        }catch (Exception e){
-            idFormaDePagoAnterior = null;
-        }
+    public void actualizar(Pago dto, String fechaComprobante) throws Exception {
+        if (dto.getFormaDePago() == null && existsByPago(dto.getId())) {
+            throw new Exception("Antes de guardar debe seleccionar una forma de pago");
+        } else {
 
-        p.setProveedor(dto.getProveedor());
-        p.setFechaComprobante(LocalDate.parse(fechaComprobante));
-        p.setTipoComprobante(dto.getTipoComprobante());
-        if(p.getTalonario() != dto.getTalonario()){
-            talonarioServicio.aumentarUltimoNro(talonarioServicio.obtenerPorNroTalonario(dto.getTalonario()));
-            p.setTalonario(dto.getTalonario());
-        }
-        p.setNroComprobante(dto.getNroComprobante());
-        p.setSector(dto.getSector());
-        p.setObservaciones(dto.getObservaciones());
-        p.setFormaDePago(dto.getFormaDePago());
-        pagoRepo.save(p);
+            Pago p = pagoRepo.findById(dto.getId()).get();
+            Long idFormaDePagoAnterior;
+            try {
+                idFormaDePagoAnterior = p.getFormaDePago().getId();
+            } catch (Exception e) {
+                idFormaDePagoAnterior = null;
+            }
 
-        if(p.getFormaDePago() != null){
-            // Si ya existe el detalleDePago verifico si cambia la formaDePago para operar
-            if(idFormaDePagoAnterior != p.getFormaDePago().getId()){
-                if(p.getFormaDePago().getId() == 54){ //La forma de pago ha cambiado y si es aDetallar => eliminar items detalleDePago
-                    formaDePagoDetalleServicio.eliminarSubDetalles(p.getId(), p.getTipoOperacion().getId());
-                }else{      // Si no es aDetallar => generar automaticamente el item detalleDepago
-                    formaDePagoDetalleServicio.eliminarSubDetalles(p.getId(), p.getTipoOperacion().getId());
-                    formaDePagoDetalleServicio.crearSubDetalleAutomatico(p.getId(), p.getTipoOperacion().getId(), p.getTotal(), p.getFormaDePago());
+            p.setProveedor(dto.getProveedor());
+            p.setFechaComprobante(LocalDate.parse(fechaComprobante));
+            p.setTipoComprobante(dto.getTipoComprobante());
+            if (p.getTalonario() != dto.getTalonario()) {
+                talonarioServicio.aumentarUltimoNro(talonarioServicio.obtenerPorNroTalonario(dto.getTalonario()));
+                p.setTalonario(dto.getTalonario());
+            }
+            p.setNroComprobante(dto.getNroComprobante());
+            p.setSector(dto.getSector());
+            p.setObservaciones(dto.getObservaciones());
+            p.setFormaDePago(dto.getFormaDePago());
+            pagoRepo.save(p);
+
+            if (p.getFormaDePago() != null) {
+                // Si ya existe el detalleDePago verifico si cambia la formaDePago para operar
+                if (idFormaDePagoAnterior != p.getFormaDePago().getId()) {
+                    if (p.getFormaDePago().getId() == 54) { //La forma de pago ha cambiado y si es aDetallar => eliminar items detalleDePago
+                        formaDePagoDetalleServicio.eliminarSubDetalles(p.getId(), p.getTipoOperacion().getId());
+                    } else {      // Si no es aDetallar => generar automaticamente el item detalleDepago
+                        formaDePagoDetalleServicio.eliminarSubDetalles(p.getId(), p.getTipoOperacion().getId());
+                        formaDePagoDetalleServicio.crearSubDetalleAutomatico(p.getId(), p.getTipoOperacion().getId(), p.getTotal(), p.getFormaDePago());
+                    }
                 }
             }
-        }
 
+        }
     }
 
+
     @Transactional(readOnly = true)
-    public Page<Pago> obtenerTodos(int page, int size){
+    public Page<Pago> obtenerTodos(int page, int size) {
         return pagoRepo.findAllByOrderByIdDesc(PageRequest.of(page, size));
     }
 
     @Transactional(readOnly = true)
-    public Pago obtenerPorId(Long id){
+    public Pago obtenerPorId(Long id) {
         return pagoRepo.findById(id).get();
     }
 
     @Transactional
-    public void eliminarPorId(Long id){
+    public void eliminarPorId(Long id) {
         pagoRepo.deleteById(id);
     }
 
     @Transactional(readOnly = true)
-    public Long buscarUltimoId() { return pagoRepo.findLastId(); }
+    public Long buscarUltimoId() {
+        return pagoRepo.findLastId();
+    }
 
     @Transactional(readOnly = true)
-    public Double obtenerTotalPorId(Long id) { return pagoRepo.obtenerTotalPorId(id);   }
+    public Double obtenerTotalPorId(Long id) {
+        return pagoRepo.obtenerTotalPorId(id);
+    }
 
 
     @Transactional
@@ -122,13 +133,13 @@ public class PagoServicio {
     }
 
     @Transactional(readOnly = true)
-    public BigDecimal obtenerTotalMensual(){
+    public BigDecimal obtenerTotalMensual() {
         return pagoRepo.obtenerTotalPagadoMensual();
     }
 
 
     @Transactional(readOnly = true)
-    public BigDecimal obtenerSaldoProveedor(Long idOperacion){
+    public BigDecimal obtenerSaldoProveedor(Long idOperacion) {
         Long fkProveedor = pagoRepo.obtenerProveedor(idOperacion);
         return proveedorServicio.obtenerSaldo(fkProveedor);
     }
@@ -139,48 +150,53 @@ public class PagoServicio {
 
     }
 
-
-    public List<PagoDTO> obtenerOperaciones(String fechaDesde, String fechaHasta, Long sectorId, Integer talDesde, Integer talHasta, Boolean excluirTal, Long idFormaPago ){
-        if(excluirTal){
+    @Transactional(readOnly = true)
+    public List<PagoDTO> obtenerOperaciones(String fechaDesde, String fechaHasta, Long sectorId, Integer talDesde, Integer talHasta, Boolean excluirTal, Long idFormaPago) {
+        if (excluirTal) {
             List<Pago> listaPagos = pagoRepo.obtenerOperacionesExcluyendoTalonario(fechaDesde, fechaHasta, sectorId, talDesde, talHasta, idFormaPago);
-            List<PagoDTO> listaPagoDTO = new ArrayList<>();
+            List<PagoDTO> listaOperacionesDTO = new ArrayList<>();
 
             for (Pago p : listaPagos) {
-                PagoDTO pagoDTO = new PagoDTO();
-                pagoDTO.setId(p.getId());
-                pagoDTO.setProveedor(entidadBaseServicio.obtenerNombrePorFkProveedor(p.getProveedor().getId()).getRazonSocial());
-                pagoDTO.setFechaComprobante(p.getFechaComprobante());
-                pagoDTO.setTalonario(p.getTalonario());
-                pagoDTO.setNroComprobante(p.getNroComprobante());
-                pagoDTO.setSector(p.getSector().getDescripcion());
-                pagoDTO.setFormaPago(p.getFormaDePago().getDescripcion());
-                pagoDTO.setTotal(p.getTotal());
-                listaPagoDTO.add(pagoDTO);
+                PagoDTO opDto = new PagoDTO();
+                opDto.setId(p.getId());
+                opDto.setEntidad(entidadBaseServicio.obtenerNombrePorFkProveedor(p.getProveedor().getId()).getRazonSocial());
+                opDto.setFechaComprobante(p.getFechaComprobante());
+                opDto.setTalonario(p.getTalonario());
+                opDto.setNroComprobante(p.getNroComprobante());
+                opDto.setSector(p.getSector().getDescripcion());
+                opDto.setFormaPago(p.getFormaDePago().getDescripcion());
+                opDto.setTotal(p.getTotal());
+                listaOperacionesDTO.add(opDto);
             }
-            return listaPagoDTO;
-        }else {
+            return listaOperacionesDTO;
+        } else {
             List<Pago> listaPagos = pagoRepo.obtenerOperaciones(fechaDesde, fechaHasta, sectorId, talDesde, talHasta, idFormaPago);
-            List<PagoDTO> listaPagoDTO = new ArrayList<>();
+            List<PagoDTO> listaOperacionesDTO = new ArrayList<>();
 
-            for (Pago p: listaPagos) {
-                PagoDTO pagoDTO = new PagoDTO();
-                pagoDTO.setId(p.getId());
-                pagoDTO.setProveedor(entidadBaseServicio.obtenerNombrePorFkProveedor(p.getProveedor().getId()).getRazonSocial());
-                pagoDTO.setFechaComprobante(p.getFechaComprobante());
-                pagoDTO.setTalonario(p.getTalonario());
-                pagoDTO.setNroComprobante(p.getNroComprobante());
-                pagoDTO.setSector(p.getSector().getDescripcion());
-                pagoDTO.setFormaPago(p.getFormaDePago().getDescripcion());
-                pagoDTO.setTotal(p.getTotal());
-                listaPagoDTO.add(pagoDTO);
+            for (Pago p : listaPagos) {
+                PagoDTO opDto = new PagoDTO();
+                opDto.setId(p.getId());
+                opDto.setEntidad(entidadBaseServicio.obtenerNombrePorFkProveedor(p.getProveedor().getId()).getRazonSocial());
+                opDto.setFechaComprobante(p.getFechaComprobante());
+                opDto.setTalonario(p.getTalonario());
+                opDto.setNroComprobante(p.getNroComprobante());
+                opDto.setSector(p.getSector().getDescripcion());
+                opDto.setFormaPago(p.getFormaDePago().getDescripcion());
+                opDto.setTotal(p.getTotal());
+                listaOperacionesDTO.add(opDto);
             }
-            return listaPagoDTO;
+            return listaOperacionesDTO;
         }
-
 
 
     }
 
 
+    @Transactional(readOnly = true)
+    public boolean  existsByPago(Long idPago){
+        if( pagoRepo.existsByPago(idPago) == 0 && pagoRepo.existsImputacionByPago(idPago) == 0) return false;
+        else return true;
+
+    }
 
 }
